skin_cancer_backend/requirements.txt
fastapi
uvicorn
python-multipart
torch
torchvision
transformers
pillow